Alice/AutomationSystem.py
import asyncio
import logging
from datetime import datetime
from typing import Dict, List, Optional, Any, Union
from dataclasses import dataclass
from enum import Enum, auto
import pyautogui
import numpy as np
import cv2
import pytesseract
from pathlib import Path

from TimePerception import QuantumTime
from ProcessManager import ProcessManager
from DataManager import DataManager

class AutomationType(Enum):
    UI = auto()
    PROCESS = auto()
    DATA = auto()
    WORKFLOW = auto()
    LEARNING = auto()

@dataclass
class AutomationTask:
    type: AutomationType
    parameters: Dict[str, Any]
    timeout: Optional[float] = None
    retries: int = 0
    quantum_sensitive: bool = False

class AutomationSystem:
    def __init__(self, quantum_time: QuantumTime):
        self.reference_time = datetime.fromisoformat("2024-12-23T03:56:26-06:00")
        self.quantum_time = quantum_time
        self.process_manager = ProcessManager(quantum_time)
        self.data_manager = DataManager()
        
        # Configure UI automation
        pyautogui.FAILSAFE = True
        pyautogui.PAUSE = 0.1
        
        # Configure OCR
        if hasattr(pytesseract, 'pytesseract'):
            pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR\tesseract.exe'
            
    async def execute_task(self, task: AutomationTask) -> Dict[str, Any]:
        """Execute an automation task with retries and timeout."""
        retries_left = task.retries
        last_error = None
        
        while retries_left >= 0:
            try:
                if task.quantum_sensitive:
                    await self._wait_for_quantum_alignment()
                    
                result = await asyncio.wait_for(
                    self._execute_task_by_type(task),
                    timeout=task.timeout
                )
                
                return {'success': True, 'result': result}
                
            except asyncio.TimeoutError:
                last_error = "Task execution timed out"
            except Exception as e:
                last_error = str(e)
                logging.error(f"Task execution failed: {last_error}")
                
            retries_left -= 1
            if retries_left >= 0:
                await asyncio.sleep(1)
                
        return {'success': False, 'error': last_error}
        
    async def _wait_for_quantum_alignment(self):
        """Wait for proper quantum phase if task is quantum-sensitive."""
        while not self.quantum_time.is_aligned():
            await asyncio.sleep(0.1)
            
    async def _execute_task_by_type(self, task: AutomationTask) -> Any:
        """Execute task based on its type."""
        if task.type == AutomationType.UI:
            return await self._execute_ui_task(task.parameters)
        elif task.type == AutomationType.PROCESS:
            return await self._execute_process_task(task.parameters)
        elif task.type == AutomationType.DATA:
            return await self._execute_data_task(task.parameters)
        elif task.type == AutomationType.WORKFLOW:
            return await self._execute_workflow_task(task.parameters)
        elif task.type == AutomationType.LEARNING:
            return await self._execute_learning_task(task.parameters)
        else:
            raise ValueError(f"Unsupported task type: {task.type}")
            
    async def _execute_ui_task(self, parameters: Dict[str, Any]) -> Any:
        """Execute UI automation task."""
        action = parameters.get('action')
        
        if action == 'click':
            if parameters['target_type'] == 'image':
                location = pyautogui.locateCenterOnScreen(
                    parameters['target_image'],
                    confidence=parameters.get('confidence', 0.9)
                )
                if location:
                    pyautogui.click(location)
                else:
                    raise ValueError(f"Image {parameters['target_image']} not found")
            elif parameters['target_type'] == 'text':
                # Use OCR to find and click text
                screenshot = pyautogui.screenshot()
                text_locations = self._find_text_locations(
                    np.array(screenshot),
                    parameters['target_text']
                )
                if text_locations:
                    x, y = text_locations[0]
                    pyautogui.click(x, y)
                else:
                    raise ValueError(f"Text {parameters['target_text']} not found")
                    
        elif action == 'type':
            pyautogui.typewrite(parameters['text'], interval=parameters.get('interval', 0.1))
            
        elif action == 'hotkey':
            pyautogui.hotkey(*parameters['keys'])
            
        return {'status': 'completed'}
        
    async def _execute_process_task(self, parameters: Dict[str, Any]) -> Any:
        """Execute process automation task."""
        return await self.process_manager.start_process(parameters['config'])
        
    async def _execute_data_task(self, parameters: Dict[str, Any]) -> Any:
        """Execute data automation task."""
        if parameters.get('operation') == 'read':
            return await self.data_manager.read_data(
                parameters['source'],
                parameters['config']
            )
        elif parameters.get('operation') == 'write':
            return await self.data_manager.write_data(
                parameters['data'],
                parameters['target'],
                parameters['config']
            )
        else:
            raise ValueError(f"Unsupported data operation: {parameters.get('operation')}")
            
    async def _execute_workflow_task(self, parameters: Dict[str, Any]) -> Any:
        """Execute workflow automation task."""
        results = []
        tasks = parameters.get('tasks', [])
        
        if parameters.get('execution_type') == 'sequential':
            for task in tasks:
                result = await self.execute_task(task)
                results.append(result)
                if not result['success'] and not parameters.get('continue_on_error', False):
                    break
        elif parameters.get('execution_type') == 'parallel':
            results = await asyncio.gather(
                *[self.execute_task(task) for task in tasks],
                return_exceptions=True
            )
        else:
            raise ValueError(f"Unsupported execution type: {parameters.get('execution_type')}")
            
        return {'results': results}
        
    async def _execute_learning_task(self, parameters: Dict[str, Any]) -> Any:
        """Execute learning automation task."""
        # Implement learning capabilities here
        return {'status': 'learning_completed'}
        
    def _find_text_locations(self, image: np.ndarray, text: str) -> List[tuple]:
        """Find locations of text in image using OCR."""
        # Convert image to grayscale
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        
        # Use pytesseract to get text locations
        data = pytesseract.image_to_data(gray, output_type=pytesseract.Output.DICT)
        
        locations = []
        for i, word in enumerate(data['text']):
            if text.lower() in word.lower():
                x = data['left'][i]
                y = data['top'][i]
                locations.append((x + data['width'][i] // 2, y + data['height'][i] // 2))
                
        return locations
        
    async def cleanup(self):
        """Clean up resources."""
        await self.process_manager.cleanup()
        await self.data_manager.cleanup()
