requirements.txt
# Core dependencies
openai>=1.3.0
anthropic>=0.3.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
psutil>=5.9.0
numpy>=1.24.0
streamlit>=1.24.0
python-dotenv>=1.0.0

# Automation dependencies
pyautogui>=0.9.54
selenium>=4.0.0

# Data processing
pandas>=2.0.0
networkx>=3.0

# Testing and development
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.0.0
flake8>=6.0.0
